0.1.45
Getting random codes based on the difficulty the user is currently in. Main part of this commit was studying the old cpp code because i needed to understand how the game worked in the first place.

Finished this commit while listening to the Furia vs Fnatic EPL in the Gaules stream.